Refactor domain layer
