--- app/src/main/java/ru/neexol/rtut/domain/models/LessonTime.kt
@@ -1,7 +1,0 @@
-package ru.neexol.rtut.domain.models
-
-@kotlinx.serialization.Serializable
-data class LessonTime(
-	val begin: String,
-	val end: String
-)
