+++ app/src/main/java/ru/neexol/rtut/presentation/grouplessons/GroupLessonsViewModel.kt
@@ -8,11 +8,11 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import ru.neexol.rtut.core.Constants
 import ru.neexol.rtut.core.Resource
-import ru.neexol.rtut.domain.models.Lesson
-import ru.neexol.rtut.domain.usecases.GetGroupLessons
-import ru.neexol.rtut.domain.usecases.GetTimes
+import ru.neexol.rtut.domain.lessons.models.DEFAULT_TIMES
+import ru.neexol.rtut.domain.lessons.models.Lesson
+import ru.neexol.rtut.domain.lessons.usecases.GetGroupLessons
+import ru.neexol.rtut.domain.lessons.usecases.GetTimes
 import javax.inject.Inject
 
 @HiltViewModel
@@ -28,7 +28,7 @@ class GroupLessonsViewModel @Inject constructor(
 	private val _lessonsResource = MutableStateFlow<Resource<List<Lesson>>>(Resource.Success(emptyList()))
 	val lessonsResource = _lessonsResource.asStateFlow()
 
-	private val _times = MutableStateFlow(Constants.DEFAULT_TIMES)
+	private val _times = MutableStateFlow(DEFAULT_TIMES)
 	val times = _times.asStateFlow()
 
 	fun loadLessons() {
