+++ app/src/main/java/ru/neexol/rtut/presentation/teacherlessons/TeacherLessonsViewModel.kt
@@ -11,11 +11,11 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import ru.neexol.rtut.core.Constants
 import ru.neexol.rtut.core.Resource
-import ru.neexol.rtut.domain.models.Lesson
-import ru.neexol.rtut.domain.usecases.GetTeacherLessons
-import ru.neexol.rtut.domain.usecases.GetTimes
+import ru.neexol.rtut.domain.lessons.models.DEFAULT_TIMES
+import ru.neexol.rtut.domain.lessons.models.Lesson
+import ru.neexol.rtut.domain.lessons.usecases.GetTeacherLessons
+import ru.neexol.rtut.domain.lessons.usecases.GetTimes
 import javax.inject.Inject
 
 @HiltViewModel
@@ -27,7 +27,7 @@ class TeacherLessonsViewModel @Inject constructor(
 	private val _lessonsResource = MutableStateFlow<Resource<List<Lesson>>>(Resource.Success(emptyList()))
 	val lessonsResource = _lessonsResource.asStateFlow()
 
-	private val _times = MutableStateFlow(Constants.DEFAULT_TIMES).apply {
+	private val _times = MutableStateFlow(DEFAULT_TIMES).apply {
 		viewModelScope.launch(Dispatchers.IO) {
 			emit(getTimesUseCase())
 		}
